requirements.txt
flask
pytest
requests
requests_mock
flake8
gitpython
